+++ packages/node/src/helpers/vimeo.ts
@@ -187,6 +187,11 @@ export default class VimeoHelper extends BaseHelper {
   }
 
   async getVideoData(videoId: string): Promise<MinimalVideoData | undefined> {
+    if (videoId.includes("?app_id=")) {
+      const [embedId] = videoId.split("?");
+      return this.returnBaseData(embedId);
+    }
+
     const isPrivate = this.isPrivatePlayer();
     if (isPrivate) {
       const videoInfo = await this.getPrivateVideoInfo(videoId);
@@ -272,6 +277,11 @@ export default class VimeoHelper extends BaseHelper {
   // eslint-disable-next-line @typescript-eslint/require-await
   async getVideoId(url: URL) {
     const embedId = /video\/[^/]+$/.exec(url.pathname)?.[0];
+    const appId = url.searchParams.get("app_id");
+    if (embedId && appId) {
+      return `${embedId}?app_id=${appId}`;
+    }
+
     if (this.isPrivatePlayer()) {
       return embedId;
     }